Rename tests annotation file
