+++ datadog_checks_base/tests/models/test_interface.py
@@ -5,7 +5,7 @@ import pytest
 from six import PY3
 
 from datadog_checks.base import AgentCheck
-from datadog_checks.dev.test_requirements import requires_py3
+from datadog_checks.dev.testing import requires_py3
 
 if PY3:
     from .config_models import ConfigMixin
