+++ datadog_checks_base/tests/openmetrics/test_interface.py
@@ -5,7 +5,7 @@ import pytest
 
 from datadog_checks.base import OpenMetricsBaseCheckV2
 from datadog_checks.base.constants import ServiceCheck
-from datadog_checks.dev.test_requirements import requires_py3
+from datadog_checks.dev.testing import requires_py3
 
 from .utils import get_check
