+++ datadog_checks_base/tests/openmetrics/test_transformers/test_native_dynamic.py
@@ -3,7 +3,8 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import pytest
 
-from ...utils import requires_py3
+from datadog_checks.dev.testing import requires_py3
+
 from ..utils import get_check
 
 pytestmark = [
