+++ datadog_checks_base/tests/openmetrics/test_transformers/test_rate.py
@@ -3,7 +3,7 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import pytest
 
-from datadog_checks.dev.test_requirements import requires_py3
+from datadog_checks.dev.testing import requires_py3
 
 from ..utils import get_check
