+++ datadog_checks_base/tests/test_agent_check.py
@@ -15,7 +15,7 @@ from datadog_checks.base import AgentCheck
 from datadog_checks.base import __version__ as base_package_version
 from datadog_checks.base import to_native_string
 from datadog_checks.base.checks.base import datadog_agent
-from datadog_checks.dev.test_requirements import requires_py3
+from datadog_checks.dev.testing import requires_py3
 
 
 def test_instance():
