+++ datadog_checks_base/tests/test_pdhbasecheck.py
@@ -5,7 +5,7 @@ import copy
 
 import pytest
 
-from datadog_checks.dev.test_requirements import requires_windows
+from datadog_checks.dev.testing import requires_windows
 
 try:
     from datadog_test_libs.win.pdh_mocks import initialize_pdh_tests, pdh_mocks_fixture  # noqa: F401
