+++ datadog_checks_base/tests/test_wmicheck.py
@@ -4,7 +4,7 @@
 
 import pytest
 
-from datadog_checks.dev.test_requirements import requires_windows
+from datadog_checks.dev.testing import requires_windows
 
 try:
     from datadog_checks.base.checks.win.wmi import WinWMICheck
