+++ datadog_checks_dev/datadog_checks/dev/testing.py
@@ -1,6 +1,9 @@
 # (C) Datadog, Inc. 2021-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+
+""" This module contains test annotations
+"""
 import pytest
 from six import PY2
